requirements.txt
httpx
beautifulsoup4
langchain-text-splitters
aiofiles
chromadb
sentence-transformers
python-dotenv
